GLOWS L1 updates (#889)
* Fixing up glows L1 tests and various small name changes/cleanup

* Most of L1B validation, and renaming things

* Fixing tests

* Adding new validition data

* minor cleanup

* type updates

* Updating CDF attributes

* fixing docs

* Updating CDF attributes

* Addressing PR comments, adjusting tests and CDF attributes

* adding conftest

* Fixing test

+++ imap_processing/glows/utils/constants.py
@@ -35,7 +35,7 @@ class TimeTuple:
         object.__setattr__(self, "seconds", final_seconds)
         object.__setattr__(self, "subseconds", final_subseconds)
 
-    def to_seconds(self) -> np.single:
+    def to_seconds(self) -> np.double:
         """
         Convert the TimeTuple to seconds.
 
@@ -44,7 +44,7 @@ class TimeTuple:
         np.single
             TimeTuple in seconds.
         """
-        return np.single(
+        return np.double(
             self.seconds + self.subseconds / GlowsConstants.SUBSECOND_LIMIT
         )
